Scope DataLayout tab split
